Adds the OWM API client
Connects the weather endpoint to the real OpenWeatherMap client to
retrieve real data.
The library JayWay JSON Path is used to map the json data received
from OWM into the internal Weather object. Despite the use of stronger
JSON parsing libraries can be used, the use of Json paths is very
handy for fast prototypes and can be used in production like
environments.

Note: the formatted sunrise and sunset times are localized to server
time.