Adds caching for calls with OWM client
Weather data from OWM doesn't refresh so often and even at their API
documentation is recommended to not issue calls more frequent than 10
minutes.

With this commit the Spring Cache feature is added to the application.
Uses Caffeine as the underlying implementation, the cache
"weatherByCityId" is created with a TTL of 10 minutes (600s)